initial dev api commit
